Use uint64 instead of int32 as IDs for aiContext
(cherry picked from commit fc2c2c93e8051083add83eee04b7edb0e894e00c)

+++ Source/abci/Importer/AlembicImporter.cpp
@@ -28,7 +28,7 @@ abciAPI void aiClearContextsWithPath(const char *path)
     aiContextManager::destroyContextsWithPath(path);
 }
 
-abciAPI aiContext* aiContextCreate(int uid)
+abciAPI aiContext* aiContextCreate(uint64_t uid)
 {
     return aiContextManager::getContext(uid);
 }
